+++ packages/core/src/v3/runMetadata/index.ts
@@ -1,8 +1,9 @@
 import { DeserializedJson } from "../../schemas/json.js";
+import { AsyncIterableStream } from "../apiClient/stream.js";
 import { getGlobal, registerGlobal } from "../utils/globals.js";
 import { ApiRequestOptions } from "../zodfetch.js";
 import { NoopRunMetadataManager } from "./noopManager.js";
-import { RunMetadataManager } from "./types.js";
+import { RunMetadataManager, RunMetadataUpdater } from "./types.js";
 
 const API_NAME = "run-metadata";
 
@@ -41,32 +42,39 @@ export class RunMetadataAPI implements RunMetadataManager {
     return this.#getManager().getKey(key);
   }
 
-  public setKey(key: string, value: DeserializedJson) {
-    return this.#getManager().setKey(key, value);
+  public set(key: string, value: DeserializedJson) {
+    this.#getManager().set(key, value);
+    return this;
   }
 
-  public deleteKey(key: string) {
-    return this.#getManager().deleteKey(key);
+  public del(key: string) {
+    this.#getManager().del(key);
+    return this;
   }
 
-  public incrementKey(key: string, value: number): void {
-    return this.#getManager().incrementKey(key, value);
+  public increment(key: string, value: number) {
+    this.#getManager().increment(key, value);
+    return this;
   }
 
-  decrementKey(key: string, value: number): void {
-    return this.#getManager().decrementKey(key, value);
+  decrement(key: string, value: number) {
+    this.#getManager().decrement(key, value);
+    return this;
   }
 
-  appendKey(key: string, value: DeserializedJson): void {
-    this.#getManager().appendKey(key, value);
+  append(key: string, value: DeserializedJson) {
+    this.#getManager().append(key, value);
+    return this;
   }
 
-  removeFromKey(key: string, value: DeserializedJson): void {
-    this.#getManager().removeFromKey(key, value);
+  remove(key: string, value: DeserializedJson) {
+    this.#getManager().remove(key, value);
+    return this;
   }
 
-  public update(metadata: Record<string, DeserializedJson>): void {
-    return this.#getManager().update(metadata);
+  public update(metadata: Record<string, DeserializedJson>) {
+    this.#getManager().update(metadata);
+    return this;
   }
 
   public stream<T>(
@@ -77,7 +85,23 @@ export class RunMetadataAPI implements RunMetadataManager {
     return this.#getManager().stream(key, value, signal);
   }
 
+  public fetchStream<T>(key: string, signal?: AbortSignal): Promise<AsyncIterableStream<T>> {
+    return this.#getManager().fetchStream(key, signal);
+  }
+
   flush(requestOptions?: ApiRequestOptions): Promise<void> {
     return this.#getManager().flush(requestOptions);
+  }
+
+  refresh(requestOptions?: ApiRequestOptions): Promise<void> {
+    return this.#getManager().refresh(requestOptions);
+  }
+
+  get parent(): RunMetadataUpdater {
+    return this.#getManager().parent;
+  }
+
+  get root(): RunMetadataUpdater {
+    return this.#getManager().root;
   }
 }